Create prework google doc
The idea for my pre work submission is based on a community garden that feeds the homeless and low income members of a neighborhood. I wanted to make an inclusive experience for those who donate or those who benefit directly from eating the meals by allowing visitors to the website to cook along from home  as meals are being prepared. This would make it feel like everyone is benefitting from a healthy meal!